fix: Export-CombinedReports now no longer requires Nessus Scans to export
Corrected the function to allow for the function to work correctly with either Nessus or CKL instead of CKL and Nessus. Also change all strings to string literals. Created Pester Tests to properly check that the function worked with either nessus or CKL instead of requiring both reports to process.